Function pmem::flush

pub fn flush<T>(x: &T)

Flushes the processor caches

This is the first of two steps in flushing to persistence, the second stpe is drain().

These steps are performed together when persist() is called.

Note that either of these steps may be unnecessary on a given platform, and the library knows how to check for that and do the right thing. For example, on Intel platforms, pmem_drain() is an empty function.