Function pmem::has_hw_drain

pub fn has_hw_drain() -> bool

Wether or not the machine supports an explicit hardware drain instruction for persistent memory.

On Intel processors with persistent memory, stores to persistent memory are considered persistent once they are flushed from the CPU caches, so this function always returns false.

Despite that, programs using flush(1) to flush ranges of memory should still follow up by calling drain()